separate spawns, dead drops, and pois into dedicated files and clean up names
