+++ src/components/map/layers/KeyMarkers.svelte
@@ -11,7 +11,7 @@
     searchTerm,
     activePanel,
   } from "$store";
-  import { keys } from "$data/key-data";
+  import { keys } from "$data/keys";
   import Markers from "../Markers.svelte";
 
   let placeMarkers, iconSettings;
