+++ src/components/map/layers/POILabels.svelte
@@ -2,7 +2,7 @@
   import L from "leaflet";
   import { onMount } from "svelte";
   import { map, layers } from "$store";
-  import { POIs } from "$data/map-data";
+  import { POIs } from "$data/pois";
 
   function placePOILabels() {
     let poiLayer = [];
