+++ src/data/dead_drops.js
@@ -1,0 +1,44 @@
+export const dead_drops = [
+  {
+    title: "C3",
+    location: {
+      x: 1203,
+      y: 1544,
+    },
+  },
+  {
+    title: "F3",
+    location: {
+      x: 2328,
+      y: 1539,
+    },
+  },
+  {
+    title: "F4",
+    location: {
+      x: 2329,
+      y: 1905,
+    },
+  },
+  {
+    title: "G6",
+    location: {
+      x: 2789,
+      y: 2507,
+    },
+  },
+  {
+    title: "E7",
+    location: {
+      x: 1684,
+      y: 2953,
+    },
+  },
+  {
+    title: "D6",
+    location: {
+      x: 1467,
+      y: 2563,
+    },
+  },
+];
