+++ src/data/pois.js
@@ -1,0 +1,121 @@
+export const POIs = [
+  {
+    title: "Taraq Village",
+    location: {
+      x: 1900,
+      y: 950,
+    },
+  },
+  {
+    title: "Rohan Oil",
+    location: {
+      x: 1480,
+      y: 1400,
+    },
+  },
+  {
+    title: "Al-Mazrah City",
+    location: {
+      x: 2450,
+      y: 1450,
+    },
+  },
+  {
+    title: "Al-Safwa Quarry",
+    location: {
+      x: 800,
+      y: 1750,
+    },
+  },
+  {
+    title: "Zarqwa Hydroelectric",
+    location: {
+      x: 1850,
+      y: 1850,
+    },
+  },
+  {
+    title: "Mawizeh Marshlands",
+    location: {
+      x: 2450,
+      y: 2000,
+    },
+  },
+  {
+    title: "Sattiq Cave Complex",
+    location: {
+      x: 1435,
+      y: 2182,
+    },
+  },
+  {
+    title: "Hafid Port",
+    location: {
+      x: 875,
+      y: 2500,
+    },
+  },
+  {
+    title: "Sa'id City",
+    location: {
+      x: 1300,
+      y: 2625,
+    },
+  },
+  {
+    title: "Zaya Observatory",
+    location: {
+      x: 2000,
+      y: 2500,
+    },
+  },
+  {
+    title: "Al Sharim Pass",
+    location: {
+      x: 2500,
+      y: 2350,
+    },
+  },
+  {
+    title: "Ahkdar Village",
+    location: {
+      x: 2300,
+      y: 2750,
+    },
+  },
+  {
+    title: "Al Samman Cemetery",
+    location: {
+      x: 1550,
+      y: 3050,
+    },
+  },
+  {
+    title: "Sawah Village",
+    location: {
+      x: 1300,
+      y: 3300,
+    },
+  },
+  {
+    title: "Sarrif Bay",
+    location: {
+      x: 2100,
+      y: 3150,
+    },
+  },
+  {
+    title: "Al Malik Airport",
+    location: {
+      x: 2850,
+      y: 2950,
+    },
+  },
+  {
+    title: "Al Bagra Fortress",
+    location: {
+      x: 2525,
+      y: 3425,
+    },
+  },
+];
