+++ src/data/spawns.js
@@ -1,170 +1,3 @@
-export const POIs = [
-  {
-    title: "Taraq Village",
-    location: {
-      x: 1900,
-      y: 950,
-    },
-  },
-  {
-    title: "Rohan Oil",
-    location: {
-      x: 1480,
-      y: 1400,
-    },
-  },
-  {
-    title: "Al-Mazrah City",
-    location: {
-      x: 2450,
-      y: 1450,
-    },
-  },
-  {
-    title: "Al-Safwa Quarry",
-    location: {
-      x: 800,
-      y: 1750,
-    },
-  },
-  {
-    title: "Zarqwa Hydroelectric",
-    location: {
-      x: 1850,
-      y: 1850,
-    },
-  },
-  {
-    title: "Mawizeh Marshlands",
-    location: {
-      x: 2450,
-      y: 2000,
-    },
-  },
-  {
-    title: "Sattiq Cave Complex",
-    location: {
-      x: 1435,
-      y: 2182,
-    },
-  },
-  {
-    title: "Hafid Port",
-    location: {
-      x: 875,
-      y: 2500,
-    },
-  },
-  {
-    title: "Sa'id City",
-    location: {
-      x: 1300,
-      y: 2625,
-    },
-  },
-  {
-    title: "Zaya Observatory",
-    location: {
-      x: 2000,
-      y: 2500,
-    },
-  },
-  {
-    title: "Al Sharim Pass",
-    location: {
-      x: 2500,
-      y: 2350,
-    },
-  },
-  {
-    title: "Ahkdar Village",
-    location: {
-      x: 2300,
-      y: 2750,
-    },
-  },
-  {
-    title: "Al Samman Cemetery",
-    location: {
-      x: 1550,
-      y: 3050,
-    },
-  },
-  {
-    title: "Sawah Village",
-    location: {
-      x: 1300,
-      y: 3300,
-    },
-  },
-  {
-    title: "Sarrif Bay",
-    location: {
-      x: 2100,
-      y: 3150,
-    },
-  },
-  {
-    title: "Al Malik Airport",
-    location: {
-      x: 2850,
-      y: 2950,
-    },
-  },
-  {
-    title: "Al Bagra Fortress",
-    location: {
-      x: 2525,
-      y: 3425,
-    },
-  },
-];
-
-export const deaddrops = [
-  {
-    title: "C3",
-    location: {
-      x: 1203,
-      y: 1544,
-    },
-  },
-  {
-    title: "F3",
-    location: {
-      x: 2328,
-      y: 1539,
-    },
-  },
-  {
-    title: "F4",
-    location: {
-      x: 2329,
-      y: 1905,
-    },
-  },
-  {
-    title: "G6",
-    location: {
-      x: 2789,
-      y: 2507,
-    },
-  },
-  {
-    title: "E7",
-    location: {
-      x: 1684,
-      y: 2953,
-    },
-  },
-  {
-    title: "D6",
-    location: {
-      x: 1467,
-      y: 2563,
-    },
-  },
-];
-
 export const spawns = [
   {
     title: "H7",
